remove the bug in project JavaSimpleCalulator, on divide method, now returms boolean(not int)

+++ JavaSimpleCalculator/src/com/step/calculator/Calculator.java
@@ -10,8 +10,8 @@ public class Calculator {
   public int multiply(int firstNumber, int secondNumber){
     return firstNumber * secondNumber;
   }
-  public int divide(int firstNumber, int secondNumber){
-    return firstNumber / secondNumber;
+  public double divide(int firstNumber, int secondNumber){
+    return (double)firstNumber / secondNumber;
   }
 
 }
